refactor: rename modules's name
